requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
